+++ Eikona/Tessa/ConnectorBundle/AttributeType/Factory/TessaValueFactory.php
@@ -2,13 +2,14 @@
 
 namespace Eikona\Tessa\ConnectorBundle\AttributeType\Factory;
 
-use Akeneo\Pim\Enrichment\Component\Product\Factory\Value\ScalarValueFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\Value\ValueFactory;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ValueInterface;
 use Akeneo\Pim\Structure\Component\Query\PublicApi\AttributeType\Attribute;
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 use Eikona\Tessa\ConnectorBundle\AttributeType\AttributeTypes;
+use Eikona\Tessa\ConnectorBundle\AttributeType\Value\TessaAssetsValue;
 
-class TessaValueFactory extends ScalarValueFactory
+class TessaValueFactory implements ValueFactory
 {
     /**
      * {@inheritDoc}
@@ -23,7 +24,34 @@ class TessaValueFactory extends ScalarValueFactory
             );
         }
 
-        return parent::createWithoutCheckingData($attribute, $channelCode, $localeCode, $data);
+        return $this->createWithoutCheckingData($attribute, $channelCode, $localeCode, $data);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function createWithoutCheckingData(
+        Attribute $attribute,
+        ?string $channelCode,
+        ?string $localeCode,
+        $data
+    ): ValueInterface {
+
+        $attributeCode = $attribute->code();
+
+        if ($attribute->isLocalizableAndScopable()) {
+            return TessaAssetsValue::scopableLocalizableValue($attributeCode, $data, $channelCode, $localeCode);
+        }
+
+        if ($attribute->isScopable()) {
+            return TessaAssetsValue::scopableValue($attributeCode, $data, $channelCode);
+        }
+
+        if ($attribute->isLocalizable()) {
+            return TessaAssetsValue::localizableValue($attributeCode, $data, $localeCode);
+        }
+
+        return TessaAssetsValue::value($attributeCode, $data);
     }
 
     public function supportedAttributeType(): string