+++ Eikona/Tessa/ReferenceDataAttributeBundle/Record/TessaValueTransformer.php
@@ -10,11 +10,26 @@ namespace Eikona\Tessa\ReferenceDataAttributeBundle\Record;
 
 use Akeneo\ReferenceEntity\Domain\Model\Attribute\AbstractAttribute;
 use Akeneo\ReferenceEntity\Infrastructure\Persistence\Sql\Record\Hydrator\Transformer\ConnectorValueTransformerInterface;
+use Eikona\Tessa\ConnectorBundle\Utilities\LinkGenerator;
 use Eikona\Tessa\ReferenceDataAttributeBundle\Attribute\TessaAttribute;
 use Webmozart\Assert\Assert;
 
 class TessaValueTransformer implements ConnectorValueTransformerInterface
 {
+    protected LinkGenerator $linkGenerator;
+
+    /**
+     * TessaValueTransformer constructor.
+     *
+     * @param LinkGenerator $linkGenerator
+     */
+    public function __construct(
+        LinkGenerator $linkGenerator
+    )
+    {
+        $this->linkGenerator = $linkGenerator;
+    }
+
     public function supports(AbstractAttribute $attribute): bool
     {
         return $attribute instanceof TessaAttribute;
@@ -24,10 +39,18 @@ class TessaValueTransformer implements ConnectorValueTransformerInterface
     {
         Assert::true($this->supports($attribute));
 
+        $data = $normalizedValue['data'];
+        $assetIds = empty($data) ? [] : $data;
+
+        $assetUrls = array_map(function($assetId) use ($normalizedValue) {
+            return $this->linkGenerator->getAssetTessaDownloadUrl($assetId, $normalizedValue['scope']);
+        }, $assetIds);
+
         return [
             'locale'  => $normalizedValue['locale'],
             'channel' => $normalizedValue['channel'],
             'data'    => $normalizedValue['data'],
+            '_links'  => $assetUrls,
         ];
     }
 }