+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * [feature] Adjust the amount of assets displayed in product grid and reference entity grid
+* [feature] Added asset download links in API
 
 ## 1.0.1 (2021-07-15)
 * [bugfix] fixed error when publishing a product